Updated Header, Carousel, HTML, and CSS elements
Updated the Header by adding an about me description to the <p>.
Updated the Carousel element by resizing the pictures and the overall size of the element on the page.
Added a padding around the skills and certifications sections.
Added an icon from Lordicon to the header in the footer.